feat: initial commit
This is the initial commit of pare a cross platform clipboard manager.
It's currently capable of taking in input via stdin or opening up a tui
where you can delete things that were copied to the clipboard or can
select it to change the clipboard to that item. Eventually I'd like it
to be able to just listen for clipboard changes automatically, but that
seems more involved. As far as day to day functionality this covers 95%
of my use cases for now.